% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BVCfit.R
\name{plot.BVCfit}
\alias{plot.BVCfit}
\title{plot a BVCfit object}
\usage{
\method{plot}{BVCfit}(x, prob=0.95, \dots)
}
\arguments{
\item{x}{BVCfit object.}

\item{prob}{probability for credible interval, between 0 and 1. e.g. prob=0.95 leads to 95\% credible interval}

\item{...}{other plot arguments}
}
\description{
plot the identified varying effects
}
\examples{
data(gExp)
spbayes=BVCfit(X, Y, Z, E, clin)
plot(spbayes)

}
\seealso{
\code{\link{BVCfit}}
}
