% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.BVCfit.R
\name{print.BVCfit.pred}
\alias{print.BVCfit.pred}
\title{print a BVCfit.pred object}
\usage{
\method{print}{BVCfit.pred}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{BVCfit object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\description{
Print a summary of a BVCfit.pred object
}
\seealso{
\code{\link{predict.BVCfit}}
}
