% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.BVCfit.R
\name{print.BVCfit}
\alias{print.BVCfit}
\title{print a BVCfit object}
\usage{
\method{print}{BVCfit}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{BVCfit object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\description{
Print a summary of a BVCfit object
}
\seealso{
\code{\link{BVCfit}}
}
