% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.BVCfit.R
\name{print.BVSelection}
\alias{print.BVSelection}
\title{print a BVSelection object}
\usage{
\method{print}{BVSelection}(x, digits = max(3, getOption("digits") - 3), \dots)
}
\arguments{
\item{x}{BVSelection object.}

\item{digits}{significant digits in printout.}

\item{...}{other print arguments}
}
\description{
Print a summary of a BVSelection object
}
\seealso{
\code{\link{BVSelection}}
}
